Find all the Yoshi behaviour tiffs in a directory hierarchy

Find all the Yoshi behaviour tiffs in a directory hierarchy

find_ybr_tiffs(path = ".", recursive = TRUE, full.names = TRUE)

Arguments

path
a character vector of full path names; the default corresponds to the working directory, getwd(). Tilde expansion (see path.expand) is performed. Missing values will be ignored. Elements with a marked encoding will be converted to the native encoding (and if that fails, considered non-existent).

recursive
logical. Should the listing recurse into directories?

full.names
a logical value. If TRUE, the directory path is prepended to the file names to give a relative file path. If FALSE, the file names (rather than paths) are returned.

Value

A data.frame containing information about the identified tiffs

See also

Examples

find_ybr_tiffs(system.file("ybr_tiffs", package='flywatch'))

#>                                                                                                                                             tiff
#> 1 /Users/runner/work/_temp/Library/flywatch/ybr_tiffs/20151130T113227_Arena1_Cam0_Mike_ST_Gr66a_20XChrimsonR/movie_Test_cam_0_measurementsXY.tif
#>    geno                time
#> 1 Gr66a 2015-11-30 11:32:27

if (FALSE) {
## Make some plots for all the experiments in a directory
tiffdf=find_ybr_tiffs("/path/to/my/flydata")
for(i in 1:nrow(tiffdf)){
  intiff=tiffdf[i,'tiff']
  outpdf=paste0(basename(dirname(intiff)),'.pdf')
  pdf(outpdf)
  plot_smoothed_displacement(intiff)
  dev.off()
  cat(".")
}
}